refactor: some refactor in response to PR feedback
